refactor: move wildcard sources to canonical package
