+++ docs/architecture/python-compatibility-shims.md
@@ -64,7 +64,7 @@ import them.
 | `storage.py` | Explicit direct re-export shim (D-08) | `easyuse_anima.infrastructure.filesystem.atomic_json` and `.paths` | #163, #186 D-08 | Existing 0.5.2 supported module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and storage compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and last-known-good/atomic-write parity |
 | `autocomplete_index.py` | Explicit direct re-export shim (D-11a) | `easyuse_anima.autocomplete.index` | #162, #186 D-11a | Existing indexed-search surface; exact seven-name `__all__` and identity fixture | External/legacy imports; `autocomplete_dataset.py` now uses the canonical owner | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and index/ranking/rebuild parity |
 | `autocomplete_dataset.py` | Partial explicit compatibility module after D-11b; prompt classification remains root-owned | `easyuse_anima.autocomplete.dataset` and `.search`; classification target waits for D-13 | #162, #186 D-11 | Existing 0.5.2 dataset/search surface canonicalized in PR #386 with direct identity aliases; final shim waits for root `anima_prompt` removal | External/legacy imports and prompt tests; `api.py` uses canonical dataset/search and root classification | Unscheduled; D-13 dependency removal, final identity surface, N+1 gate, and classification/result/API parity |
-| `wildcard_engine.py` | Partial compatibility module after D-12a; source/snapshot/selector/seed/expansion remain root-owned | `easyuse_anima.wildcard.models`, then sources/snapshot/expansion modules | #184, #186 D-12 | Existing 0.5.2 immutable model/limit surface canonicalized in PR #387 with direct identity aliases; later D-12 slices remain | root entrypoint, `nodes.py`, `api.py`, wildcard/workflow tests | Unscheduled; full D-12 move, final identity surface, N+1 gate, and seed/expansion/workflow parity |
+| `wildcard_engine.py` | Partial compatibility module after D-12b; snapshot/selector/seed/expansion remain root-owned | `easyuse_anima.wildcard.models`, `.sources`, then snapshot/expansion modules | #184, #186 D-12 | Existing 0.5.2 model/limit surface canonicalized in PR #387 and source/path surface in PR #388 with direct identity aliases; later D-12 slices remain | root entrypoint, `nodes.py`, `api.py`, wildcard/workflow tests | Unscheduled; full D-12 move, final identity surface, N+1 gate, and seed/expansion/workflow parity |
 | `prompt_translation.py` | Explicit direct re-export shim (D-01) | `easyuse_anima.translation.*` | #164, #186 D-01 | Existing 0.5.2 supported module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and translation compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and provider-off/API parity |
 | `anima_prompt/` package | Current implementation; planned package shim | `easyuse_anima.prompt.anima.*` | #184, #186 D-13 | Existing 0.5.2 surface; convert in D-13/D-14 | `nodes.py`, `autocomplete_dataset.py`, prompt tests | Unscheduled; N+1 gate and prompt correction/parser parity |
 
@@ -1464,9 +1464,13 @@ EasyUseAnimaWildcard
 - D-12a moves immutable `WildcardOption`, `WildcardExpansionBudget`,
   `WildcardExpansionResult`, and their fixed budget constants to
   `easyuse_anima.wildcard.models`.
-- Root binds the 12 moved supported names directly to canonical objects.
-  Source scanning, snapshot cache/condition, selector/seed, parsing, expansion,
-  enforcement, and diagnostics remain root-owned for later D-12 slices.
+- D-12b moves root discovery, extra-path resolution, TXT/YAML parsing,
+  immutable source metadata, and source scanning to
+  `easyuse_anima.wildcard.sources`.
+- Root binds the 12 model names and eight supported source names directly to
+  canonical objects. Snapshot construction/cache/condition/single-flight,
+  selector/seed, expansion, enforcement, and diagnostics remain root-owned for
+  later D-12 slices.
 - Canonical target for the remaining implementation:
   `easyuse_anima.wildcard` sources/snapshot/expansion modules. Snapshot
   lifecycle/factory/cleanup remains E-06.
