+++ tests/test_registry_scanner_safety.py
@@ -83,6 +83,7 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/settings/service.py",
     "easyuse_anima/wildcard/__init__.py",
     "easyuse_anima/wildcard/models.py",
+    "easyuse_anima/wildcard/sources.py",
     "easyuse_anima/registration.py",
     "easyuse_anima/prompt/__init__.py",
     "easyuse_anima/prompt/correction.py",
