+++ wildcard_engine.py
@@ -5,7 +5,6 @@ from collections import OrderedDict
 import fnmatch
 import hashlib
 import math
-import os
 import random
 import re
 import threading
@@ -17,16 +16,6 @@ from typing import Iterable, Mapping, Sequence
 # NumPy is mandatory in supported ComfyUI runtimes and defines the seeded
 # wildcard sampling contract. A stdlib fallback would produce different results.
 import numpy as np
-
-try:
-    import yaml
-except Exception:  # pragma: no cover - YAML files are skipped without PyYAML.
-    yaml = None
-
-try:
-    from .easyuse_anima.infrastructure.filesystem.paths import USER_DATA_DIR
-except ImportError:
-    from easyuse_anima.infrastructure.filesystem.paths import USER_DATA_DIR
 
 try:
     from .easyuse_anima.wildcard.models import (
@@ -59,10 +48,36 @@ except ImportError:
         WildcardOption,
     )
 
+try:
+    from .easyuse_anima.wildcard import sources as _wildcard_sources
+    from .easyuse_anima.wildcard.sources import (
+        DEFAULT_TEST_WILDCARD_FILE,
+        DEFAULT_TEST_WILDCARD_TEXT,
+        WILDCARD_DIR_NAME,
+        WILDCARD_EXTENSIONS,
+        _WildcardSourceFile,
+        _WildcardSourceState,
+        default_wildcard_root,
+        ensure_default_wildcard_root,
+        parse_wildcard_extra_paths,
+        resolve_wildcard_roots,
+    )
+except ImportError:
+    from easyuse_anima.wildcard import sources as _wildcard_sources
+    from easyuse_anima.wildcard.sources import (
+        DEFAULT_TEST_WILDCARD_FILE,
+        DEFAULT_TEST_WILDCARD_TEXT,
+        WILDCARD_DIR_NAME,
+        WILDCARD_EXTENSIONS,
+        _WildcardSourceFile,
+        _WildcardSourceState,
+        default_wildcard_root,
+        ensure_default_wildcard_root,
+        parse_wildcard_extra_paths,
+        resolve_wildcard_roots,
+    )
 
-WILDCARD_DIR_NAME = "wildcards"
-DEFAULT_TEST_WILDCARD_FILE = "easyuse_anima_test.txt"
-DEFAULT_TEST_WILDCARD_TEXT = "# EasyUse Anima test wildcard\nsimple wildcard\nanima wildcard\n"
+
 WILDCARD_MODE_POPULATE = "populate"
 WILDCARD_MODE_FIXED = "fixed"
 WILDCARD_MODE_SEQUENTIAL = "sequential"
@@ -107,8 +122,6 @@ SEED_CONTROL_MODES = (
 
 MAX_SEED = 0xFFFFFFFFFFFFFFFF
 PUBLIC_MAX_SEED = (1 << 53) - 1
-WILDCARD_EXTENSIONS = {".txt", ".yaml", ".yml"}
-
 COMMENT_RE = re.compile(r"^\s*#.*(?:\n|$)", re.MULTILINE)
 DYNAMIC_RE = re.compile(r"(?<![\\%])\{((?:[^{}]|(?<=\\)[{}])*?)(?<!\\)\}")
 WILDCARD_RE = re.compile(r"__(?P<keyword>[\w.\-+/*\\]+?)__", re.IGNORECASE)
@@ -117,39 +130,9 @@ WILDCARD_QUANTIFIER_RE = re.compile(
     r"(?P<quantifier>\d+)#__(?P<keyword>[\w.\-+/*\\]+?)__",
     re.IGNORECASE,
 )
-WEIGHT_PREFIX_RE = re.compile(r"^\s*([+-]?(?:\d+(?:\.\d*)?|\.\d+))::(.*)$", re.DOTALL)
 COUNT_SPEC_RE = re.compile(
     r"(?:(?P<fixed>\d+)|(?P<minimum>\d*)\s*-\s*(?P<maximum>\d*))"
 )
-
-
-@dataclass(frozen=True)
-class _WildcardSourceFile:
-    root_index: int
-    root: str
-    relative_path: str
-    path: Path
-    mtime_ns: int
-    size: int
-
-    @property
-    def cache_key(self) -> tuple[int, str, int, int]:
-        return (self.root_index, self.relative_path, self.mtime_ns, self.size)
-
-
-@dataclass(frozen=True)
-class _WildcardSourceState:
-    roots: tuple[Path, ...]
-    root_identities: tuple[str, ...]
-    files: tuple[_WildcardSourceFile, ...]
-
-    @property
-    def cache_key(self) -> tuple:
-        roots = tuple(
-            (identity, str(root))
-            for identity, root in zip(self.root_identities, self.roots)
-        )
-        return roots, tuple(source.cache_key for source in self.files)
 
 
 @dataclass(frozen=True)
@@ -322,7 +305,9 @@ class _ExpansionState:
             cycle_parts = (*candidate.parts, candidate.separator)
         for part in cycle_parts:
             for match in WILDCARD_RE.finditer(part):
-                key = _normalize_wildcard_key(match.group("keyword"))
+                key = _wildcard_sources._normalize_wildcard_key(
+                    match.group("keyword")
+                )
                 if key is not None and key in key_stack:
                     self.cycle_detected = True
                     return None
@@ -405,20 +390,6 @@ class _ExpansionState:
         return _ExpansionText(output)
 
 
-def default_wildcard_root() -> Path:
-    return USER_DATA_DIR / WILDCARD_DIR_NAME
-
-
-def ensure_default_wildcard_root(create_sample: bool = True) -> Path:
-    root = default_wildcard_root()
-    root.mkdir(parents=True, exist_ok=True)
-    if create_sample:
-        sample_path = root / DEFAULT_TEST_WILDCARD_FILE
-        if not sample_path.exists():
-            sample_path.write_text(DEFAULT_TEST_WILDCARD_TEXT, encoding="utf-8")
-    return root
-
-
 def normalize_wildcard_mode(mode: str) -> str:
     value = str(mode or "").strip()
     return WILDCARD_MODE_ALIASES.get(value, WILDCARD_MODE_POPULATE)
@@ -469,231 +440,13 @@ def has_wildcard_syntax(text: str) -> bool:
     return bool(DYNAMIC_RE.search(value) or WILDCARD_RE.search(value) or WILDCARD_QUANTIFIER_RE.search(value))
 
 
-def parse_wildcard_extra_paths(value: str) -> list[str]:
-    paths = []
-    for line in str(value or "").splitlines():
-        path = line.strip().strip('"')
-        if path:
-            paths.append(path)
-    return paths
-
-
-def _comfy_base_path() -> Path:
-    try:
-        import folder_paths  # type: ignore
-
-        base_path = getattr(folder_paths, "base_path", None)
-        if base_path:
-            return Path(base_path)
-    except Exception:
-        pass
-    return Path.cwd()
-
-
-def _resolve_path(value: str, base_path: Path) -> Path:
-    expanded = os.path.expandvars(os.path.expanduser(value))
-    path = Path(expanded)
-    if not path.is_absolute():
-        path = base_path / path
-    return path.resolve()
-
-
-def resolve_wildcard_roots(extra_paths: str | None = None) -> list[Path]:
-    if extra_paths is None:
-        try:
-            from .easyuse_anima.settings.repository import get_settings
-        except ImportError:
-            from easyuse_anima.settings.repository import get_settings
-
-        extra_paths = get_settings().get("wildcard.extra_paths", "")
-
-    base_path = _comfy_base_path()
-    roots: list[Path] = []
-    seen = set()
-    for raw_path in parse_wildcard_extra_paths(extra_paths or ""):
-        try:
-            root = _resolve_path(raw_path, base_path)
-        except (OSError, RuntimeError, ValueError):
-            continue
-        key = os.path.normcase(str(root))
-        if key not in seen:
-            seen.add(key)
-            roots.append(root)
-
-    try:
-        default_root = ensure_default_wildcard_root().resolve()
-    except OSError:
-        default_root = default_wildcard_root().resolve()
-    default_key = os.path.normcase(str(default_root))
-    if default_key not in seen:
-        roots.append(default_root)
-    return roots
-
-
-def _normalize_wildcard_key(value: str) -> str | None:
-    key = str(value or "").replace("\\", "/").replace(" ", "-").strip().strip("/")
-    if not key:
-        return None
-    if key.startswith("/") or re.match(r"^[a-zA-Z]:", key):
-        return None
-    parts = [part for part in key.split("/") if part]
-    if any(part == ".." for part in parts):
-        return None
-    return "/".join(parts).lower()
-
-
-def _read_text_file(path: Path) -> str:
-    try:
-        return path.read_text(encoding="utf-8")
-    except UnicodeDecodeError:
-        return path.read_text(encoding="iso-8859-1")
-
-
-def _parse_option(value) -> WildcardOption | None:
-    text = "" if value is None else str(value).strip()
-    if not text:
-        return None
-    match = WEIGHT_PREFIX_RE.match(text)
-    if not match:
-        return WildcardOption(text=text, weight=1.0)
-    try:
-        weight = float(match.group(1))
-    except ValueError:
-        return WildcardOption(text=text, weight=1.0)
-    return WildcardOption(text=match.group(2).strip(), weight=max(0.0, weight))
-
-
-def _options_from_lines(text: str) -> list[WildcardOption]:
-    options = []
-    for raw_line in text.splitlines():
-        line = raw_line.strip()
-        if not line or line.startswith("#"):
-            continue
-        option = _parse_option(line)
-        if option is not None:
-            options.append(option)
-    return options
-
-
-def _stringify_yaml_scalar(value) -> str:
-    if isinstance(value, bool):
-        return "true" if value else "false"
-    return "" if value is None else str(value)
-
-
-def _yaml_entries(data, prefix: str = "") -> dict[str, list[WildcardOption]]:
-    entries: dict[str, list[WildcardOption]] = {}
-
-    def collect(value, path_prefix: str, publish_alias: bool = True) -> list[WildcardOption]:
-        aggregate: list[WildcardOption] = []
-        if isinstance(value, dict):
-            for raw_key, child_value in value.items():
-                child_key = _normalize_wildcard_key(raw_key)
-                if child_key is None:
-                    continue
-                path_key = f"{path_prefix}/{child_key}" if path_prefix else child_key
-                aggregate.extend(collect(child_value, path_key))
-        elif isinstance(value, list):
-            for item in value:
-                if isinstance(item, (dict, list)):
-                    # The containing list owns this alias. Nested containers still
-                    # publish their distinct child paths, but not the same prefix
-                    # again through an intermediate recursion frame.
-                    aggregate.extend(collect(item, path_prefix, publish_alias=False))
-                    continue
-                option = _parse_option(_stringify_yaml_scalar(item))
-                if option is not None:
-                    aggregate.append(option)
-        else:
-            option = _parse_option(_stringify_yaml_scalar(value))
-            if option is not None:
-                aggregate.append(option)
-
-        if publish_alias and path_prefix and aggregate:
-            entries.setdefault(path_prefix, []).extend(aggregate)
-        return aggregate
-
-    collect(data, prefix)
-    return entries
-
-
-def _load_yaml_entries(path: Path) -> dict[str, list[WildcardOption]]:
-    if yaml is None:
-        return {}
-    text = _read_text_file(path)
-    try:
-        data = yaml.safe_load(text)
-    except Exception:
-        return {}
-    return _yaml_entries(data)
-
-
-def _load_wildcard_file(root: Path, path: Path) -> dict[str, list[WildcardOption]]:
-    suffix = path.suffix.lower()
-    if suffix not in WILDCARD_EXTENSIONS:
-        return {}
-    if suffix in {".yaml", ".yml"}:
-        return _load_yaml_entries(path)
-    try:
-        relative_key = path.relative_to(root).with_suffix("").as_posix()
-    except ValueError:
-        return {}
-    key = _normalize_wildcard_key(relative_key)
-    if key is None:
-        return {}
-    return {key: _options_from_lines(_read_text_file(path))}
-
-
-def _wildcard_root_identity(root: Path) -> str:
-    try:
-        return os.path.normcase(os.path.abspath(os.fspath(root)))
-    except (OSError, TypeError, ValueError):
-        return os.path.normcase(str(root))
-
-
-def _scan_wildcard_sources(roots: tuple[Path, ...]) -> _WildcardSourceState:
-    files: list[_WildcardSourceFile] = []
-    for root_index, root in enumerate(roots):
-        if not root.is_dir():
-            continue
-        try:
-            candidates = sorted(root.rglob("*"), key=lambda item: item.as_posix().lower())
-        except OSError:
-            continue
-        for path in candidates:
-            if path.suffix.lower() not in WILDCARD_EXTENSIONS:
-                continue
-            try:
-                if not path.is_file():
-                    continue
-                stat = path.stat()
-                relative = path.relative_to(root).as_posix()
-            except (OSError, ValueError):
-                continue
-            files.append(
-                _WildcardSourceFile(
-                    root_index=root_index,
-                    root=str(root),
-                    relative_path=relative,
-                    path=path,
-                    mtime_ns=stat.st_mtime_ns,
-                    size=stat.st_size,
-                )
-            )
-    return _WildcardSourceState(
-        roots=roots,
-        root_identities=tuple(_wildcard_root_identity(root) for root in roots),
-        files=tuple(files),
-    )
-
-
 def _build_wildcard_snapshot(source_state: _WildcardSourceState) -> _WildcardSnapshot:
     mapping: dict[str, list[WildcardOption]] = {}
     cacheable = True
     for source in source_state.files:
         root = source_state.roots[source.root_index]
         try:
-            entries = _load_wildcard_file(root, source.path)
+            entries = _wildcard_sources._load_wildcard_file(root, source.path)
         except OSError:
             cacheable = False
             continue
@@ -717,7 +470,7 @@ def _build_wildcard_snapshot(source_state: _WildcardSourceState) -> _WildcardSna
 def _wildcard_snapshot(roots: Iterable[Path]) -> _WildcardSnapshot:
     resolved_roots = tuple(Path(root) for root in roots)
     while True:
-        source_state = _scan_wildcard_sources(resolved_roots)
+        source_state = _wildcard_sources._scan_wildcard_sources(resolved_roots)
         cache_key = source_state.cache_key
         with _SNAPSHOT_CONDITION:
             cached = _SNAPSHOT_CACHE.get(cache_key)
@@ -733,7 +486,7 @@ def _wildcard_snapshot(roots: Iterable[Path]) -> _WildcardSnapshot:
         failure: BaseException | None = None
         try:
             candidate = _build_wildcard_snapshot(source_state)
-            verified_state = _scan_wildcard_sources(resolved_roots)
+            verified_state = _wildcard_sources._scan_wildcard_sources(resolved_roots)
             if verified_state.cache_key == cache_key:
                 snapshot = candidate
         except BaseException as exc:
@@ -852,7 +605,7 @@ class _WildcardLibrary:
             self.missing.append(key)
 
     def options_for(self, raw_key: str) -> Sequence[WildcardOption]:
-        key = _normalize_wildcard_key(raw_key)
+        key = _wildcard_sources._normalize_wildcard_key(raw_key)
         if key is None:
             return []
         options = self._options_for_normalized_key(key)
@@ -908,7 +661,7 @@ def _split_unescaped(value: str, separator: str) -> list[str]:
 def _parse_dynamic_options(value: str) -> list[WildcardOption]:
     options = []
     for option in _split_unescaped(value, "|"):
-        parsed = _parse_option(option)
+        parsed = _wildcard_sources._parse_option(option)
         if parsed is not None:
             options.append(parsed)
     return options
@@ -943,7 +696,10 @@ def _expand_multiselect_options(
     if not match:
         return options, None
     raw_key = match.group("keyword")
-    return library.options_for(raw_key), _normalize_wildcard_key(raw_key)
+    return (
+        library.options_for(raw_key),
+        _wildcard_sources._normalize_wildcard_key(raw_key),
+    )
 
 
 def _replace_dynamic(
@@ -1003,7 +759,7 @@ def _replace_quantified_wildcards(
         raw_key = match.group("keyword")
         options = library.options_for(raw_key)
         selected = selector.choose_many(options, count)
-        key = _normalize_wildcard_key(raw_key)
+        key = _wildcard_sources._normalize_wildcard_key(raw_key)
         if not selected or key is None:
             return None
         return state.candidate(
@@ -1025,7 +781,7 @@ def _replace_file_wildcards(
         raw_key = match.group("keyword")
         options = library.options_for(raw_key)
         selected = selector.choose_one(options)
-        key = _normalize_wildcard_key(raw_key)
+        key = _wildcard_sources._normalize_wildcard_key(raw_key)
         if selected is None or key is None:
             return None
         return state.candidate((selected.text,), "", key_stack + (key,))
